Move layout_test sub-module out of builtin
